requirements.txt
opencv-python
pycocotools
pyyaml
tensorboard
torchvision==0.5.0
tqdm
webcolors
